Restructured files, refactored App component, and externalized some styles to their related style documents.
